+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Require an additional `--already-transferred` flag for the single-message form of `quill neuron-stake`. (#173)
 - Added `--disburse-amount` and `--disburse-to` to `quill neuron-manage`. (#171)
 - Accepts bare principals and ICRC-1 account IDs in `quill account-balance` and `quill transfer`. (#168)
 - Allowed omitting the account ID in `quill account-balance`. (#167)